Migrate replace-node to 'api' module
