+++ solr/core/src/java/org/apache/solr/handler/admin/api/ReplaceNode.java
@@ -16,7 +16,6 @@
  */
 package org.apache.solr.handler.admin.api;
 
-import static org.apache.solr.client.solrj.impl.BinaryResponseParser.BINARY_CONTENT_TYPE_V2;
 import static org.apache.solr.cloud.Overseer.QUEUE_OPERATION;
 import static org.apache.solr.common.params.CollectionParams.SOURCE_NODE;
 import static org.apache.solr.common.params.CollectionParams.TARGET_NODE;
@@ -25,17 +24,11 @@ import static org.apache.solr.common.params.CommonAdminParams.WAIT_FOR_FINAL_STA
 import static org.apache.solr.handler.admin.CollectionsHandler.DEFAULT_COLLECTION_OP_TIMEOUT;
 import static org.apache.solr.security.PermissionNameProvider.Name.COLL_EDIT_PERM;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
-import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.media.Schema;
-import io.swagger.v3.oas.annotations.parameters.RequestBody;
 import java.util.HashMap;
 import java.util.Map;
 import javax.inject.Inject;
-import javax.ws.rs.POST;
-import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
-import javax.ws.rs.Produces;
+import org.apache.solr.client.api.endpoint.ReplaceNodeApi;
+import org.apache.solr.client.api.model.ReplaceNodeRequestBody;
 import org.apache.solr.client.api.model.SolrJerseyResponse;
 import org.apache.solr.client.solrj.SolrResponse;
 import org.apache.solr.common.cloud.ZkNodeProps;
@@ -43,7 +36,6 @@ import org.apache.solr.common.params.CollectionParams;
 import org.apache.solr.common.params.CollectionParams.CollectionAction;
 import org.apache.solr.core.CoreContainer;
 import org.apache.solr.handler.admin.CollectionsHandler;
-import org.apache.solr.jersey.JacksonReflectMapWriter;
 import org.apache.solr.jersey.PermissionName;
 import org.apache.solr.request.SolrQueryRequest;
 import org.apache.solr.response.SolrQueryResponse;
@@ -53,26 +45,19 @@ import org.apache.solr.response.SolrQueryResponse;
  *
  * <p>This API is analogous to the v1 /admin/collections?action=REPLACENODE command.
  */
-@Path("cluster/nodes/{sourceNodeName}/replace/")
-public class ReplaceNodeAPI extends AdminAPIBase {
+public class ReplaceNode extends AdminAPIBase implements ReplaceNodeApi {
 
   @Inject
-  public ReplaceNodeAPI(
+  public ReplaceNode(
       CoreContainer coreContainer,
       SolrQueryRequest solrQueryRequest,
       SolrQueryResponse solrQueryResponse) {
     super(coreContainer, solrQueryRequest, solrQueryResponse);
   }
 
-  @POST
-  @Produces({"application/json", "application/xml", BINARY_CONTENT_TYPE_V2})
+  @Override
   @PermissionName(COLL_EDIT_PERM)
-  public SolrJerseyResponse replaceNode(
-      @Parameter(description = "The name of the node to be replaced.", required = true)
-          @PathParam("sourceNodeName")
-          String sourceNodeName,
-      @RequestBody(description = "Contains user provided parameters", required = true)
-          ReplaceNodeRequestBody requestBody)
+  public SolrJerseyResponse replaceNode(String sourceNodeName, ReplaceNodeRequestBody requestBody)
       throws Exception {
     final SolrJerseyResponse response = instantiateJerseyResponse(SolrJerseyResponse.class);
     final CoreContainer coreContainer = fetchAndValidateZooKeeperAwareCoreContainer();
@@ -110,35 +95,5 @@ public class ReplaceNodeAPI extends AdminAPIBase {
     if (value != null) {
       dest.put(key, value);
     }
-  }
-
-  public static class ReplaceNodeRequestBody implements JacksonReflectMapWriter {
-
-    public ReplaceNodeRequestBody() {}
-
-    public ReplaceNodeRequestBody(String targetNodeName, Boolean waitForFinalState, String async) {
-      this.targetNodeName = targetNodeName;
-      this.waitForFinalState = waitForFinalState;
-      this.async = async;
-    }
-
-    @Schema(
-        description =
-            "The target node where replicas will be copied. If this parameter is not provided, Solr "
-                + "will identify nodes automatically based on policies or number of cores in each node.")
-    @JsonProperty("targetNodeName")
-    public String targetNodeName;
-
-    @Schema(
-        description =
-            "If true, the request will complete only when all affected replicas become active. "
-                + "If false, the API will return the status of the single action, which may be "
-                + "before the new replica is online and active.")
-    @JsonProperty("waitForFinalState")
-    public Boolean waitForFinalState = false;
-
-    @Schema(description = "Request ID to track this action which will be processed asynchronously.")
-    @JsonProperty("async")
-    public String async;
   }
 }
